+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 ### Changes
 
 - Improve patient search performance
+- Improve loading recent patients performance
 - Show patient died status for dead patients in search results
 - [In Progress: 22 Feb 2022] Add stronger app update reminders